Setup Redux store
