+++ src/index.js
@@ -1,14 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { Provider } from 'react-redux';
 import './styles/index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { store } from './redux/configureStore';
 import axios from 'axios';
+
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <App />
+    <Provider store={store}>
+      <App />
+    </Provider>
   </React.StrictMode>
 );
