+++ src/redux/configureStore.js
@@ -1,0 +1,9 @@
+import { continentSlice } from "./continentsSlice";
+import { configureStore } from "@reduxjs/toolkit/";
+
+export const store = configureStore({
+  reducer: {
+    continents: continentSlice.reducer
+  }
+})
+
